# Convert the remaining functions in `src/core/primitives.js` to use standard classes